[Add] UIComponent 파일 분리 (#127)
